requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
gitdb==4.0.9
GitPython==3.1.26
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
smmap==5.0.0
Werkzeug==2.0.2
